Added supervisor SysV init script template
When running the buildout `etc/sysv-init.in` creates `etc/sysv-init` which acts
as a System V init script for the supervisor daemon used to run the adhocracy
services.

To use it, link it to `/etc/init.d/adhocracy-adhocracy.lan` or alike and enable
it by linking the file to `/etc/rc2.d/`. This is in the same style as
`vhost.conf.in` for the apache virtual host configuration.

The script is pretty much a copy of the adhocracy supervisord init script
`etc/init.d__adhocracy_services.sh.template` by phihag.

Differences:

 - It uses buildout to fill in buildout directory and user name under which
   supervisord will run, while the other script uses `sed` from within
   `build_debian.sh`.

 - It uses `sudo` instead of `su` to change the user.

 - The name appends the used domain in order to allow a setup with multiple
   adhocracy installations.